+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## v1.5.31 (Draft)
+## v1.6.0 (Draft)
 
-### 🪄 Improvements
+### ✨ New
 
 - Add default settings for `react-x` settings to presets.
+- Undeprecate rule `no-implicit-key` and improve its usefulness.
+- Undeprecate rule `no-complicated-conditional-rendering` and rename it to `no-complex-conditional-rendering` (the previous rule will still be available until the next major update to avoid breaking changes).
 
 ## v1.5.30 (Mon 22 Jul 2024)